After review: add tests
